调整编译grub的脚本的部分 (#108)
1、bugfix: 修复编译grub的脚本的部分错误
2、将grub下载源替换为tuna
3、优化写入磁盘镜像的脚本
4、将bios文件夹改名为legacy

+++ tools/grub_auto_install.sh
@@ -1,12 +1,15 @@
 #!/bin/bash
 grub_dir_i386_efi=arch/i386/efi/grub
-grub_dir_i386_bios=arch/i386/bios/grub
+grub_dir_i386_legacy=arch/i386/legacy/grub
 grub_dir_x86_64_efi=arch/x86_64/efi/grub
-#编译核心数目
-nproc=8
+
+mkdir -p ${grub_dir_i386_efi}
+mkdir -p ${grub_dir_i386_legacy}
+mkdir -p ${grub_dir_x86_64_efi}
+
 
 #检测grub是否已经安装
-if [ -d ${grub_dir_i386_efi}/bin ] || [ -d ${grub_dir_i386_bios}/bin ] || [ -d ${grub_dir_x86_64_efi}/bin ] ; then
+if [ -d ${grub_dir_i386_efi}/bin ] || [ -d ${grub_dir_i386_legacy}/bin ] && [ -d ${grub_dir_x86_64_efi}/bin ] ; then
 	exit 0
 fi
 #仅支持Ubuntu/Debain下的自动安装
@@ -15,45 +18,49 @@ if ! hash 2>/dev/null apt-get; then
 	exit 0
 fi
 #下载grub2.06
+if [ ! -f "grub-2.06.tar.xz" ]; then
+    wget https://mirrors.tuna.tsinghua.edu.cn/gnu/grub/grub-2.06.tar.xz || exit 1
+fi
+
 echo "开始下载grub2.06"
-wget https://ftp.gnu.org/gnu/grub/grub-2.06.tar.gz
+
 echo "下载完成"
-tar -xvzf grub-2.06.tar.gz
+tar xvf grub-2.06.tar.xz
 #安装对应依赖
 sudo apt-get update
 sudo apt-get install -y \
-	make \
+	make 	\
   	binutils \
-  	bison \
-  	gcc \
+  	bison 	\
+  	gcc 	\
   	gettext \
-	flex\
-	bison\
-	automake\
-	autoconf\	
+	flex	\
+	bison	\
+	automake	\
+	autoconf	
 	
 cd grub-2.06
 echo "开始安装grub2.06"
 #编译安装三个版本的grub
-./configure --target=i386 --prefix=$(dirname $PWD)/${grub_dir_i386_bios}
-make -j $(nporc)
-make install
-make clean 
+./configure --target=i386 --prefix=$(dirname $PWD)/${grub_dir_i386_legacy} || exit 1
+make -j $(nproc) || exit 1
+make install || exit 1
+make clean || exit 1
 
-./configure --target=i386 --with-platform=efi --prefix=$(dirname $PWD)/${grub_dir_i386_efi}
-make -j $(nporc)
-make install
-make clean 
+./configure --target=i386 --with-platform=efi --prefix=$(dirname $PWD)/${grub_dir_i386_efi} ||	exit 1
+make -j $(nproc) || exit 1
+make install || exit 1
+make clean || exit 1
 
-./configure --target=x86_64 --with-platform=efi --prefix=$(dirname $PWD)/${grub_dir_x86_64_efi}
-make -j $(nporc)
-make install
+./configure --target=x86_64 --with-platform=efi --prefix=$(dirname $PWD)/${grub_dir_x86_64_efi} || exit 1
+make -j $(nproc) || exit 1
+make install || exit 1
 
 cd ..
 #解除权限限制
-sudo chmod -R 777 ${grub_dir_i386_bios}
+sudo chmod -R 777 ${grub_dir_i386_legacy}
 sudo chmod -R 777 ${grub_dir_i386_efi}
 sudo chmod -R 777 ${grub_dir_x86_64_efi}
 rm -rf grub-2.06
-rm grub-2.06.tar.gz
+rm grub-2.06.tar.xz*
 echo "grub2.06安装完成"
